# openapi.yaml
openapi: 3.0.4
info:
  title: Code RAG API
  description: >-
    API for managing indexed code projects and querying their vectorized source
    code using natural language. Projects support full CRUD via POST/GET/PUT/DELETE
    /projects. The service stores per-project code documents (functions, types, members, etc.)
    along with vector embeddings, and allows retrieving the most semantically similar
    pieces of code for a given natural language question via a pgvector-backed similarity search.


    All endpoints exclusively accept and return application/json. Client errors
    and server errors are reported using the RFC 7807 "Problem Details for HTTP APIs"
    format, with the exception of 404 responses, which are returned with no response body.
  version: 1.0.0
paths:
...
  /api/v1/projects:
    get:
      tags:
        - Projects
      summary: List all projects
      description: >-
        Returns all registered projects. Optionally filter the results by
        project name using a

        partial, case-insensitive match against the name field. Use the returned
        id values as the

        projectId path parameter when fetching, renaming, or deleting a specific
        project, or when

        querying code via the /projects/{projectId}/code-queries endpoint.
                    
        When no project matches the supplied filter (or no projects exist at
        all), the response is

        a 200 OK with an empty array - this is not treated as an error.
      responses:
        "200":
          description: >-
            A list of projects matching the given filter, or all projects if no
            filter was supplied.

            Returns an empty array when there are no matches.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectResponse"
        "400":
          description: >-
            The name query parameter is invalid (e.g. it exceeds the maximum
            allowed length, or is

            present but empty). This is the only condition under which this
            endpoint returns 400.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: >-
            An unhandled exception occurred while processing the request. This
            is the only condition

            under which this endpoint returns 500.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ServerErrorProblemDetails"
    post:
      tags:
        - Projects
      summary: Create a project
      description: >-
        Registers a new project so its source code can later be indexed and
        searched. The

        project's name must be unique across all projects (case-sensitive, exact
        match); creating

        a project with a name that already exists returns a 409. git_url and
        git_raw_url are

        optional and, when provided, are validated but not otherwise checked for
        reachability.
                    
        On success, the response includes a Location header pointing to the new
        project's

        GET /projects/{projectId} URL, per the standard HTTP convention for 201
        Created responses.
      requestBody:
        description: The new project's name and, optionally, its git repository details.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectRequest"
      responses:
        "201":
          description: The project was created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400":
          description: >-
            The request body is missing or malformed, the name field is
            missing/empty/blank, name

            exceeds the maximum allowed length, or git_url/git_raw_url was
            provided but is blank or

            exceeds the maximum allowed length. This is the only client-error
            condition other than 409

            under which this endpoint fails.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: >-
            A project with the given name already exists. This is the only
            condition under which this

            endpoint returns 409.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: >-
            An unhandled exception occurred while processing the request. This
            is the only condition

            under which this endpoint returns 500.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ServerErrorProblemDetails"
  /api/v1/projects/{projectId}:
    get:
      tags:
        - Projects
      summary: Get a project by id
      description: Returns a single project by its id.
      operationId: GetProject
      parameters:
        - name: projectId
          in: path
          description: >-
            Identifier of the project, corresponding to the id field returned by
            GET /projects or

            POST /projects. Must be a positive 64-bit integer; any other format
            (e.g. a GUID or

            non-numeric string) results in a 400 response.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The project matching projectId.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400":
          description: >-
            The projectId path parameter is not a valid positive integer. This
            is the only condition

            under which this endpoint returns 400.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: >-
            No project exists with the given projectId. This is the only
            condition under which this

            endpoint returns 404; the response has no body.
        "500":
          description: >-
            An unhandled exception occurred while processing the request. This
            is the only condition

            under which this endpoint returns 500.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ServerErrorProblemDetails"
    put:
      tags:
        - Projects
      summary: Rename a project
      description: >-
        Replaces the mutable fields of an existing project: name, git_url, and
        git_raw_url. This is

        a full replace, not a partial patch - fields omitted from the request
        body are cleared (set

        to null), so a client that only wants to rename a project must resend
        its current

        git_url/git_raw_url alongside the new name (e.g. from a prior GET
        /projects/{projectId}).

        name is subject to the same uniqueness constraint as POST /projects
        (renaming a project to

        its own current name is allowed and is a no-op). id and created_at are
        immutable and cannot

        be changed.
      parameters:
        - name: projectId
          in: path
          description: >-
            Identifier of the project, corresponding to the id field returned by
            GET /projects or

            POST /projects. Must be a positive 64-bit integer; any other format
            (e.g. a GUID or

            non-numeric string) results in a 400 response.
          required: true
          schema:
            type: string
      requestBody:
        description: The project's new name and git repository details.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectRequest"
      responses:
        "200":
          description: The project was updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400":
          description: >-
            Either the projectId path parameter is not a valid positive integer,
            or the request body is

            missing or malformed, or the name field is missing/empty/blank, or
            name exceeds the maximum

            allowed length, or git_url/git_raw_url was provided but is blank or
            exceeds the maximum

            allowed length. This is the only client-error condition other than
            404/409 under which this

            endpoint fails.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: >-
            No project exists with the given projectId. This is the only
            condition under which this

            endpoint returns 404; the response has no body.
        "409":
          description: >-
            Another project already has the given name. This is the only
            condition under which this

            endpoint returns 409.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: >-
            An unhandled exception occurred while processing the request. This
            is the only condition

            under which this endpoint returns 500.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ServerErrorProblemDetails"
    delete:
      tags:
        - Projects
      summary: Delete a project
      description: >-
        Permanently deletes a project. To prevent silently orphaning or
        destroying indexed code, a

        project that still has indexed code documents cannot be deleted -
        remove/re-index those

        first (outside the scope of this API). This is a hard delete; there is
        no soft-delete or

        undo.
      parameters:
        - name: projectId
          in: path
          description: >-
            Identifier of the project, corresponding to the id field returned by
            GET /projects or

            POST /projects. Must be a positive 64-bit integer; any other format
            (e.g. a GUID or

            non-numeric string) results in a 400 response.
          required: true
          schema:
            type: string
      responses:
        "204":
          description: The project was deleted. The response has no body.
        "400":
          description: >-
            The projectId path parameter is not a valid positive integer. This
            is the only

            client-error condition other than 404/409 under which this endpoint
            fails.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: >-
            No project exists with the given projectId. This is the only
            condition under which this

            endpoint returns 404; the response has no body.
        "409":
          description: >-
            The project has one or more indexed code documents and cannot be
            deleted. This is the only

            condition under which this endpoint returns 409.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: >-
            An unhandled exception occurred while processing the request. This
            is the only condition

            under which this endpoint returns 500.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ServerErrorProblemDetails"
components:
  schemas:
...
    ExceptionDetails:
      type: object
      properties:
        exception_type:
          type: string
          description: Fully-qualified type name of the exception that was raised.
          nullable: true
        message:
          type: string
          description: The exception's message.
          nullable: true
        stack_trace:
          type: string
          description: The exception's captured stack trace.
          nullable: true
      additionalProperties: false
      description: Details of the unhandled exception that caused the request to fail.
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: {}
    ProjectRequest:
      type: object
      properties:
        name:
          type: string
          description: >-
            The project's name. Must be unique across all projects
            (case-sensitive, exact match) and must

            not be empty or blank. Attempting to create or rename a project to a
            name that is already in

            use results in a 409.
          nullable: true
        git_url:
          type: string
          description: >-
            URL of the project's git repository. Optional; omit or send null for
            a project with no known

            repository URL. When provided, must not be blank and must be at most
            2000 characters long.
          nullable: true
        git_raw_url:
          type: string
          description: >-
            Base URL for fetching raw file contents from the project's git
            repository. Optional; omit or

            send null when not applicable. When provided, must not be blank and
            must be at most 2000

            characters long.
          nullable: true
      additionalProperties: false
      description: >-
        Request body for creating (`POST /projects`) or replacing (`PUT
        /projects/{projectId}`)

        a project. For `PUT`, this is a full replace: omitting
        `git_url`/`git_raw_url`

        clears them (sets them to null) rather than leaving the project's
        current values untouched.
    ProjectResponse:
      type: object
      properties:
        id:
          type: integer
          description: >-
            Auto-generated, unique identifier of the project. Used as the
            `projectId` path parameter

            when querying code for this project.
          format: int64
        name:
          type: string
          description: >-
            Unique, human-readable name of the project. Used as the searchable
            field for

            `GET /projects?name=`, and as the field set by `POST /projects` and

            `PUT /projects/{projectId}`.
          nullable: true
        git_url:
          type: string
          description: >-
            URL of the project's git repository, e.g. its clone/remote URL.
            Optional; null when not set.

            Not validated for reachability - only for being non-blank and within
            the maximum allowed

            length when provided.
          nullable: true
        git_raw_url:
          type: string
          description: >-
            Base URL for fetching raw file contents from the project's git
            repository, e.g. a "raw"

            content host URL that a client can append a matched code document's
            source file path to in

            order to link directly to the matched source. Optional; null when
            not set. Not validated for

            reachability - only for being non-blank and within the maximum
            allowed length when provided.
          nullable: true
        created_at:
          type: string
          description: >-
            Timestamp (UTC) at which the project record was created. Defaults to
            the insertion time on

            the database if not otherwise specified.
          format: date-time
      additionalProperties: false
      description: >-
        A project whose source code has been analyzed and indexed for semantic
        search. Serializes as

        snake_case.
    ServerErrorProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
        exception:
          $ref: "#/components/schemas/ExceptionDetails"
      additionalProperties: {}
      description: >-
        Problem Details payload for 500 responses. Exposing the raised
        exception's details is a

        deliberate trade-off for ease of debugging in this API, per the OpenAPI
        contract, and should

        be disabled or redacted in a hardened deployment.
...